Add authentication and authorization checks in admin.js and security.js
