+++ contact.html
@@ -1,74 +1,120 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
+  <head>
+    <meta charset="UTF-8" />
     <title>Kromann Reuter</title>
-    <link rel="stylesheet" href="./css/main.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-          integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
+    <link rel="stylesheet" href="./css/main.css" />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
+      crossorigin="anonymous"
+    />
     <script src="js/main.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-            integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-            crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-</head>
-<body>
-
-<nav class="navbar main-color text-light">
-    <div class="container-fluid d-flex justify-content-between align-items-center">
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+      crossorigin="anonymous"
+    ></script>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+    />
+  </head>
+  <body>
+    <nav class="navbar main-color text-light">
+      <div
+        class="container-fluid d-flex justify-content-between align-items-center"
+      >
         <a class="navbar-brand" href="events.html">
-            <img src="assets/images/kromann-logo-white.svg" alt="Logo" width="300" class="d-inline-block align-text-top">
+          <img
+            src="assets/images/kromann-logo-white.svg"
+            alt="Logo"
+            width="300"
+            class="d-inline-block align-text-top"
+          />
         </a>
         <div class="d-flex align-items-center">
-            <a href="events.html" class="nav-link me-5">Kommende begivenheder <i class="fa fa-calendar"></i></a>
-            <a href="login.html" class="nav-link me-2">Login <i class="fa fa-user"></i></a>
+          <a href="events.html" class="nav-link me-5"
+            >Kommende begivenheder <i class="fa fa-calendar"></i
+          ></a>
+          <a href="login.html" class="nav-link me-2" id="loginLink"
+            >Login <i class="fa fa-user"></i
+          ></a>
         </div>
-    </div>
-</nav>
+      </div>
+    </nav>
 
-<div class="container-fluid p-0 secondary-color">
-    <div class="content-container">
+    <div class="container-fluid p-0 secondary-color">
+      <div class="content-container">
         <h3>Kontakt Kromann Reuter</h3>
+      </div>
+
+      <div class="container">
+        <div class="row pb-5 justify-content-center">
+          <div class="col-lg-8">
+            <div class="box border main-color p-5" style="color: white">
+              <h3>Kontakt os:</h3>
+              <form class="mb-4">
+                <div class="form-group">
+                  <label for="name">Navn:</label>
+                  <input
+                    type="text"
+                    class="form-control"
+                    id="name"
+                    placeholder=""
+                  />
+                </div>
+                <div class="form-group">
+                  <label for="email">Email-addresse:</label>
+                  <input
+                    type="email"
+                    class="form-control"
+                    id="email"
+                    placeholder="example@mail.com"
+                  />
+                </div>
+                <div class="form-group">
+                  <label for="message">Besked:</label>
+                  <textarea
+                    class="form-control"
+                    id="message"
+                    rows="5"
+                    placeholder=""
+                  ></textarea>
+                </div>
+              </form>
+              <a href="events.html" class="primary-btn"
+                >Send <i class="fa fa-chevron-right"></i
+              ></a>
+            </div>
+          </div>
+        </div>
+      </div>
     </div>
 
     <div class="container">
-        <div class="row pb-5 justify-content-center">
-            <div class="col-lg-8">
-                <div class="box border main-color p-5" style="color: white;">
-
-                        <h3>Kontakt os:</h3>
-                        <form class="mb-4">
-                            <div class="form-group">
-                                <label for="name">Navn:</label>
-                                <input type="text" class="form-control" id="name" placeholder="">
-                            </div>
-                            <div class="form-group">
-                                <label for="email">Email-addresse:</label>
-                                <input type="email" class="form-control" id="email" placeholder="example@mail.com">
-                            </div>
-                            <div class="form-group">
-                                <label for="message">Besked:</label>
-                                <textarea class="form-control" id="message" rows="5" placeholder=""></textarea>
-                            </div>
-                        </form>
-                    <a href="events.html" class="primary-btn">Send <i class="fa fa-chevron-right"></i></a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-
-<div class="container">
-    <footer class="py-3 my-4 default_cursor_cs">
-        <ul class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs">
-            <li class="nav-item"><a href="index.html" class="nav-link px-2 text-muted">Hjem</a></li>
-            <li class="nav-item"><a href="about.html" class="nav-link px-2 text-muted">Om os</a></li>
-            <li class="nav-item"><a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a></li>
-            <li class="nav-item"><a href="login.html" class="nav-link px-2 text-muted">Log ind</a></li>
+      <footer class="py-3 my-4 default_cursor_cs">
+        <ul
+          class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs"
+        >
+          <li class="nav-item">
+            <a href="index.html" class="nav-link px-2 text-muted">Hjem</a>
+          </li>
+          <li class="nav-item">
+            <a href="about.html" class="nav-link px-2 text-muted">Om os</a>
+          </li>
+          <li class="nav-item">
+            <a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a>
+          </li>
+          <li class="nav-item">
+            <a href="login.html" id="loginLink" class="nav-link px-2 text-muted"
+              >Log ind</a
+            >
+          </li>
         </ul>
         <p class="text-center text-muted">© 2023 KEA</p>
-    </footer>
-</div>
-
-</body>
+      </footer>
+    </div>
+  </body>
 </html>
