+++ event.html
@@ -22,68 +22,103 @@
     />
   </head>
   <body>
-  <nav class="navbar main-color text-light">
-    <div class="container-fluid d-flex justify-content-between align-items-center">
-      <a class="navbar-brand" href="events.html">
-        <img src="assets/images/kromann-logo-white.svg" alt="Logo" width="300" class="d-inline-block align-text-top">
-      </a>
-      <div class="d-flex align-items-center">
-        <a href="events.html" class="nav-link me-5">Kommende begivenheder <i class="fa fa-calendar"></i></a>
-        <a href="admin.html" class="nav-link me-2">Login <i class="fa fa-user"></i></a>
+    <nav class="navbar main-color text-light">
+      <div
+        class="container-fluid d-flex justify-content-between align-items-center"
+      >
+        <a class="navbar-brand" href="events.html">
+          <img
+            src="assets/images/kromann-logo-white.svg"
+            alt="Logo"
+            width="300"
+            class="d-inline-block align-text-top"
+          />
+        </a>
+        <div class="d-flex align-items-center">
+          <a href="events.html" class="nav-link me-5"
+            >Kommende begivenheder <i class="fa fa-calendar"></i
+          ></a>
+          <a href="admin.html" class="nav-link me-2" id="loginLink"
+            >Login <i class="fa fa-user"></i
+          ></a>
+        </div>
+      </div>
+    </nav>
+
+    <div
+      class="container-fluid secondary-color d-flex justify-content-center align-items-center"
+    >
+      <div class="content-container">
+        <h3 id="event-title"></h3>
+        <img src="" alt="" id="event-image" />
+        <p id="event-date"></p>
+        <p id="event-description"></p>
+        <h3 id="form-title"></h3>
+
+        <form id="eventAnswer" class="d-flex flex-column">
+          <div class="form-check">
+            <input
+              class="form-check-input event-answer"
+              type="radio"
+              name="eventAnswer"
+              id="radioAnswer1"
+              checked
+            />
+            <label
+              class="form-check-label event-answer-label"
+              for="radioAnswer1"
+            >
+              Jeg Deltager
+            </label>
+          </div>
+          <div class="form-check">
+            <input
+              class="form-check-input event-answer"
+              type="radio"
+              name="eventAnswer"
+              id="radioAnswer2"
+            />
+            <label
+              class="form-check-label event-answer-label"
+              for="radioAnswer2"
+            >
+              Jeg Deltager Ikke
+            </label>
+          </div>
+          <button
+            type="submit"
+            id="eventAnswer-btn"
+            class="btn btn-primary mt-3"
+          >
+            Afgiv svar
+          </button>
+        </form>
       </div>
     </div>
-  </nav>
 
-  <div class="container-fluid secondary-color d-flex justify-content-center align-items-center">
-    <div class="content-container">
-      <h3 id="event-title"></h3>
-      <img src="" alt="" id="event-image" />
-      <p id="event-date"></p>
-      <p id="event-description"></p>
-      <h3 id="form-title"></h3>
-
-      <form id="eventAnswer" class="d-flex flex-column">
-        <div class="form-check">
-          <input
-                  class="form-check-input event-answer"
-                  type="radio"
-                  name="eventAnswer"
-                  id="radioAnswer1"
-                  checked
-          />
-          <label class="form-check-label event-answer-label" for="radioAnswer1">
-            Jeg Deltager
-          </label>
-        </div>
-        <div class="form-check">
-          <input
-                  class="form-check-input event-answer"
-                  type="radio"
-                  name="eventAnswer"
-                  id="radioAnswer2"
-          />
-          <label class="form-check-label event-answer-label" for="radioAnswer2">
-            Jeg Deltager Ikke
-          </label>
-        </div>
-        <button type="submit" id="eventAnswer-btn" class="btn btn-primary mt-3">
-          Afgiv svar
-        </button>
-      </form>
+    <div class="container">
+      <footer class="py-3 my-4 default_cursor_cs">
+        <ul
+          class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs"
+        >
+          <li class="nav-item">
+            <a href="index.html" class="nav-link px-2 text-muted">Hjem</a>
+          </li>
+          <li class="nav-item">
+            <a href="about.html" class="nav-link px-2 text-muted">Om os</a>
+          </li>
+          <li class="nav-item">
+            <a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a>
+          </li>
+          <li class="nav-item">
+            <a href="login.html" id="loginLink" class="nav-link px-2 text-muted"
+              >Log ind</a
+            >
+          </li>
+        </ul>
+        <p class="text-center text-muted">© 2023 KEA</p>
+      </footer>
     </div>
-  </div>
-
-  <div class="container">
-    <footer class="py-3 my-4 default_cursor_cs">
-      <ul class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs">
-        <li class="nav-item"><a href="index.html" class="nav-link px-2 text-muted">Hjem</a></li>
-        <li class="nav-item"><a href="about.html" class="nav-link px-2 text-muted">Om os</a></li>
-        <li class="nav-item"><a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a></li>
-        <li class="nav-item"><a href="login.html" class="nav-link px-2 text-muted">Log ind</a></li>
-      </ul>
-      <p class="text-center text-muted">© 2023 KEA</p>
-    </footer>
-  </div>
 
     <script src="js/event.js"></script>
   </body>
