+++ events.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <title>Kromann Reuter</title>
+    <script src="js/auth.js" type="module"></script>
     <link rel="stylesheet" href="./css/main.css" />
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
@@ -22,17 +23,28 @@
     />
   </head>
   <body>
-  <nav class="navbar main-color text-light">
-      <div class="container-fluid d-flex justify-content-between align-items-center">
-          <a class="navbar-brand" href="index.html">
-              <img src="assets/images/kromann-logo-white.svg" alt="Logo" width="300" class="d-inline-block align-text-top">
-          </a>
-          <div class="d-flex align-items-center">
-              <a href="events.html" class="nav-link me-5">Kommende begivenheder <i class="fa fa-calendar"></i></a>
-              <a href="admin.html" class="nav-link me-2">Login <i class="fa fa-user"></i></a>
-          </div>
+    <nav class="navbar main-color text-light">
+      <div
+        class="container-fluid d-flex justify-content-between align-items-center"
+      >
+        <a class="navbar-brand" href="index.html">
+          <img
+            src="assets/images/kromann-logo-white.svg"
+            alt="Logo"
+            width="300"
+            class="d-inline-block align-text-top"
+          />
+        </a>
+        <div class="d-flex align-items-center">
+          <a href="events.html" class="nav-link me-5"
+            >Kommende begivenheder <i class="fa fa-calendar"></i
+          ></a>
+          <a href="admin.html" class="nav-link me-2" id="loginLink"
+            >Login <i class="fa fa-user"></i
+          ></a>
+        </div>
       </div>
-  </nav>
+    </nav>
 
     <div class="container-fluid p-0 secondary-color">
       <div class="content-container">
@@ -42,20 +54,32 @@
       <div class="container">
         <div class="row justify-content-center" id="cardsRow"></div>
       </div>
-        <ul class="pagination d-flex justify-content-center" id="pagination"></ul>
+      <ul class="pagination d-flex justify-content-center" id="pagination"></ul>
     </div>
 
-  <div class="container">
-    <footer class="py-3 my-4 default_cursor_cs">
-      <ul class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs">
-        <li class="nav-item"><a href="index.html" class="nav-link px-2 text-muted">Hjem</a></li>
-        <li class="nav-item"><a href="about.html" class="nav-link px-2 text-muted">Om os</a></li>
-        <li class="nav-item"><a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a></li>
-        <li class="nav-item"><a href="login.html" class="nav-link px-2 text-muted">Log ind</a></li>
-      </ul>
-      <p class="text-center text-muted">© 2023 KEA</p>
-    </footer>
-  </div>
+    <div class="container">
+      <footer class="py-3 my-4 default_cursor_cs">
+        <ul
+          class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs"
+        >
+          <li class="nav-item">
+            <a href="index.html" class="nav-link px-2 text-muted">Hjem</a>
+          </li>
+          <li class="nav-item">
+            <a href="about.html" class="nav-link px-2 text-muted">Om os</a>
+          </li>
+          <li class="nav-item">
+            <a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a>
+          </li>
+          <li class="nav-item">
+            <a href="login.html" id="loginLink" class="nav-link px-2 text-muted"
+              >Log ind</a
+            >
+          </li>
+        </ul>
+        <p class="text-center text-muted">© 2023 KEA</p>
+      </footer>
+    </div>
 
     <script src="js/events.js"></script>
   </body>
