+++ login.html
@@ -38,7 +38,7 @@
           <a href="events.html" class="nav-link me-5"
             >Kommende begivenheder <i class="fa fa-calendar"></i
           ></a>
-          <a href="login.html" class="nav-link me-2"
+          <a href="login.html" class="nav-link me-2" id="loginLink"
             >Login <i class="fa fa-user"></i
           ></a>
         </div>
@@ -90,7 +90,9 @@
             <a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a>
           </li>
           <li class="nav-item">
-            <a href="login.html" class="nav-link px-2 text-muted">Log ind</a>
+            <a href="login.html" id="loginLink" class="nav-link px-2 text-muted"
+              >Log ind</a
+            >
           </li>
         </ul>
         <p class="text-center text-muted">© 2023 KEA</p>
