+++ template.html
@@ -1,109 +1,135 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Kromann Reuter</title>
-  <link rel="stylesheet" href="./css/main.css">
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-  <script src="js/main.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-          integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-          crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-</head>
-<body>
+  <head>
+    <meta charset="UTF-8" />
+    <title>Kromann Reuter</title>
+    <link rel="stylesheet" href="./css/main.css" />
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
+      crossorigin="anonymous"
+    />
+    <script src="js/main.js"></script>
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+      crossorigin="anonymous"
+    ></script>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css"
+    />
+  </head>
+  <body>
+    <nav class="navbar main-color text-light">
+      <div
+        class="container-fluid d-flex justify-content-between align-items-center"
+      >
+        <a class="navbar-brand" href="events.html">
+          <img
+            src="assets/images/kromann-logo-white.svg"
+            alt="Logo"
+            width="300"
+            class="d-inline-block align-text-top"
+          />
+        </a>
+        <div class="d-flex align-items-center">
+          <a href="events.html" class="nav-link me-5"
+            >Kommende begivenheder <i class="fa fa-calendar"></i
+          ></a>
+          <a href="login.html" class="nav-link me-2" id="loginLink"
+            >Login <i class="fa fa-user"></i
+          ></a>
+        </div>
+      </div>
+    </nav>
 
-<nav class="navbar main-color text-light">
-  <div class="container-fluid d-flex justify-content-between align-items-center">
-    <a class="navbar-brand" href="events.html">
-      <img src="assets/images/kromann-logo-white.svg" alt="Logo" width="300" class="d-inline-block align-text-top">
-    </a>
-    <div class="d-flex align-items-center">
-      <a href="events.html" class="nav-link me-5">Kommende begivenheder <i class="fa fa-calendar"></i></a>
-      <a href="login.html" class="nav-link me-2">Login <i class="fa fa-user"></i></a>
-    </div>
-  </div>
-</nav>
-
-<div class="container-fluid p-0 secondary-color">
+    <div class="container-fluid p-0 secondary-color">
       <div class="content-container">
-      <h3>Template</h3>
+        <h3>Template</h3>
       </div>
 
-<div class="container">
-  <div class="row">
-    <div class="col-md-6">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Noget her</h3>
-          <p>Bla bla bla bla</p>
+      <div class="container">
+        <div class="row">
+          <div class="col-md-6">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
+          <div class="col-md-6">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Og noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <div class="row mt-5 justify-content-center">
+          <div class="col-lg-10">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Eller noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <div class="row mt-5 pb-5">
+          <div class="col">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
+          <div class="col">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
+          <div class="col">
+            <div class="box border main-color">
+              <div class="content-container">
+                <h3>Noget her</h3>
+                <p>Bla bla bla bla</p>
+              </div>
+            </div>
+          </div>
         </div>
       </div>
     </div>
-    <div class="col-md-6">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Og noget her</h3>
-          <p>Bla bla bla bla</p>
-        </div>
-      </div>
+
+    <div class="container">
+      <footer class="py-3 my-4 default_cursor_cs">
+        <ul
+          class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs"
+        >
+          <li class="nav-item">
+            <a href="index.html" class="nav-link px-2 text-muted">Hjem</a>
+          </li>
+          <li class="nav-item">
+            <a href="about.html" class="nav-link px-2 text-muted">Om os</a>
+          </li>
+          <li class="nav-item">
+            <a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a>
+          </li>
+          <li class="nav-item">
+            <a href="login.html" class="nav-link px-2 text-muted">Log ind</a>
+          </li>
+        </ul>
+        <p class="text-center text-muted">© 2023 KEA</p>
+      </footer>
     </div>
-  </div>
-
-  <div class="row mt-5 justify-content-center">
-    <div class="col-lg-10 ">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Eller noget her</h3>
-          <p>Bla bla bla bla</p>
-        </div>
-    </div>
-  </div>
-</div>
-
-  <div class="row mt-5 pb-5">
-    <div class="col">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Noget her</h3>
-          <p>Bla bla bla bla</p>
-        </div>
-      </div>
-    </div>
-    <div class="col">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Noget her</h3>
-          <p>Bla bla bla bla</p>
-        </div>
-      </div>
-    </div>
-    <div class="col">
-      <div class="box border main-color">
-        <div class="content-container">
-          <h3>Noget her</h3>
-          <p>Bla bla bla bla</p>
-        </div>
-      </div>
-    </div>
-  </div>
-
-</div>
-</div>
-
-
-<div class="container">
-  <footer class="py-3 my-4 default_cursor_cs">
-    <ul class="nav justify-content-center border-bottom pb-3 mb-3 default_cursor_cs">
-      <li class="nav-item"><a href="index.html" class="nav-link px-2 text-muted">Hjem</a></li>
-      <li class="nav-item"><a href="about.html" class="nav-link px-2 text-muted">Om os</a></li>
-      <li class="nav-item"><a href="contact.html" class="nav-link px-2 text-muted">Kontakt</a></li>
-      <li class="nav-item"><a href="login.html" class="nav-link px-2 text-muted">Log ind</a></li>
-    </ul>
-    <p class="text-center text-muted">© 2023 KEA</p>
-  </footer>
-</div>
-
-</body>
+  </body>
 </html>
